Minimum Height Trees
For a undirected graph with tree characteristics, we can choose any node
as the root. The result graph is then a rooted tree. Among all possible
rooted trees, those with minimum height are called minimum height trees
(MHTs). Given such a graph, write a function to find all the MHTs and
return a list of their root labels.